Create folder Introduction-to-Programming-2020
The purpose of this folder is to store my solutions to the tasks from the Introduction to Programming Workshop, led by assistant Yosif Ivanov. The exposition of the study material by topics, type and conditions of the tasks with the presented sample input and output are presented in his repository - github.com/ymivanov/Introduction-to-Programming-2020

Thank you for the opportunity to upload my solutions with requirements of your problems, Yosif!